+++ include/Engine2D/Rendering/Renderer2D.hpp
@@ -26,6 +26,7 @@ using Engine::Rendering::Shader;
 
 namespace Engine2D::Rendering {
   class SpriteRenderer;
+  using Flush = std::tuple<uint32_t, uint32_t, uint32_t, uint32_t>;
 
   /// Class responsible for rendering all the textures/sprites in the game
   class Renderer2D {
@@ -35,10 +36,14 @@ namespace Engine2D::Rendering {
     /// The vertices that will be sent to the VAO
     static const float vertices[];
 
-    /// All the sprite renderers in the scene
-    static std::vector<std::shared_ptr<SpriteRenderer>> renderers;
-    /// All the static sprite renderers in the scene
-    static std::vector<std::shared_ptr<SpriteRenderer>> staticRenderers;
+    /// All the sprite renderers in the scene that are opaque
+    static std::vector<std::shared_ptr<SpriteRenderer>> opaqueRenderers;
+    /// All the sprite renderers in the scene that are transparent
+    static std::vector<std::shared_ptr<SpriteRenderer>> transparentRenderers;
+    /// All the static sprite renderers in the scene that are opaque
+    static std::vector<std::shared_ptr<SpriteRenderer>> opaqueStaticRenderers;
+    /// All the static sprite renderers in the scene that are transparent
+    static std::vector<std::shared_ptr<SpriteRenderer>> transparentStaticRenderers;
 
     /// All the sprite renderers pending to be added
     static std::vector<std::shared_ptr<SpriteRenderer>> renderersToAdd;
@@ -60,17 +65,22 @@ namespace Engine2D::Rendering {
     static uint lastShaderID;
     /// The last texture id that was used
     static uint lastTextureID;
+    static size_t lastOpaqueCount;
+    static size_t lastTransparentCount;
 
     /// Data linked to all the renderers in the scene
     static std::vector<float> batchData;
-    /// Data linked to all the renderers in the scene
-    static std::vector<float> staticBatchData;
-    /// Tells the renderer how to use the static batch data
-    static std::vector<std::tuple<uint32_t, uint32_t, uint32_t, uint32_t>> staticFlushList;
+    /// Data linked to all the renderers in the scene for opaque sprites
+    static std::vector<float> opaqueStaticBatchData;
+    /// Tells the renderer how to use the static batch data for opaque sprites
+    static std::vector<Flush> opaqueStaticFlushList;
+    /// Data linked to all the renderers in the scene for transparent sprites
+    static std::vector<float> transparentStaticBatchData;
+    /// Tells the renderer how to use the static batch data for transparent sprites
+    static std::vector<Flush> transparentStaticFlushList;
 
     Renderer2D();
     ~Renderer2D();
-
     /// Initializes the vertex data and buffers for rendering a quad.
     static void initRenderData();
     /// @return True if the renderer can be rendered, false if not
@@ -78,16 +88,24 @@ namespace Engine2D::Rendering {
     /// Returns true if a needs to be before b, False if not
     static bool sortRenderers(const std::shared_ptr<SpriteRenderer> &a, const std::shared_ptr<SpriteRenderer> &b);
     /// Extracts all the data needed to render the sprite renderer
-    static void extractRendererData(const std::shared_ptr<SpriteRenderer> &renderer, float *data);
+    static void extractRendererData(const std::shared_ptr<SpriteRenderer> &renderer, float *data, bool transparent);
 
     /// Updates the data linked to the stored static sprites
-    static void updateStaticBatch();
+    static void updateStaticBatch(
+      const std::vector<std::shared_ptr<SpriteRenderer>> &renderers, std::vector<float> &staticBatchData, bool transparent
+    );
     /// Builds and renders the dynamic batch
-    static void buildAndRenderStaticBatch(bool rebuild);
+    static void buildAndRenderStaticBatch(
+      std::vector<std::shared_ptr<SpriteRenderer>> &renderers, std::vector<float> &staticBatchData,
+      std::vector<Flush> &staticFlushList, bool transparent, uint lastStaticCount
+    );
     /// Builds and renders the dynamic batch
-    static void buildAndRenderDynamicBatch();
+    static void buildAndRenderDynamicBatch(const std::vector<std::shared_ptr<SpriteRenderer>> &renderers, bool transparent);
+
+    /// Updates the sprite lists
+    static void prerender();
     /// Renders all the given entities
-    static void render();
+    static void render(bool renderTransparent);
     /// Sends all the data to the GPU
     static void flush(uint VBO, const float *data, int drawMode, uint32_t count, uint32_t framebuffer = 0);