+++ Changelog.md
@@ -3,7 +3,16 @@
 All notable changes to this project will be documented in this file.<br>
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [0.7.2] - 2025-04-dd
+## [0.7.3] - 2025-04-27
+### Added
+- **[Sprite]**: transparent flag
+### Changed
+- **[ResourceManager]**: when loading or creating a sprite the transparent flag is needed, and rect optional
+- **[Renderer2D]**: split rendering into opaque then transparent
+### Fixed
+- **[Renderer2D]**: particles not rendering correctly under transparent sprites
+
+## [0.7.2] - 2025-04-27
 ### Added
 - **[Animation2D]**: creates a horizontal or vertical animation from a sprite sheet
 - **[Animator2D]**: manages a list of Animation2D and updates them accordingly